+++ pkg/chaining/registry.go
@@ -6,19 +6,26 @@ import (
 	"github.com/TurnsCoffeeIntoScripts/jira-api-issue-resource/pkg/noop"
 	"github.com/TurnsCoffeeIntoScripts/jira-api-issue-resource/pkg/reading"
 	"github.com/TurnsCoffeeIntoScripts/jira-api-issue-resource/pkg/service"
+	"github.com/TurnsCoffeeIntoScripts/jira-api-issue-resource/pkg/status"
 )
 
 const (
+	ServiceFetchIssueData      = "srv_fetch_issue_data"
 	ServiceReadIssueName       = "srv_read_issue"
 	ServiceEditCustomFieldName = "srv_edit_field"
+	ServiceGetTransitions      = "srv_get_transitions"
+	ServiceDoTransition        = "srv_do_transitions"
 	ServiceUnknownName         = "srv_unknown"
 )
 
 var serviceRegistry = make(map[string]service.Service)
 
 func InitServiceRegistry() {
+	serviceRegistry[ServiceFetchIssueData] = &reading.ServiceFetchIssueData{}
 	serviceRegistry[ServiceReadIssueName] = &reading.ServiceReadIssue{}
 	serviceRegistry[ServiceEditCustomFieldName] = &editing.ServiceEditCustomField{}
+	serviceRegistry[ServiceGetTransitions] = &status.ServiceGetTransitions{}
+	serviceRegistry[ServiceDoTransition] = &status.ServiceDoTransition{}
 	serviceRegistry[ServiceUnknownName] = &noop.ServiceUnknown{}
 }