+++ pkg/helpers/results_chaining_keys.go
@@ -1,6 +1,8 @@
 package helpers
 
 const (
-	ReadingFieldKey = "FieldKey"
-	ParentIssueKey  = "ParentIssueKey"
+	ReadingFieldKey   = "FieldKey"          //
+	ParentIssueKey    = "ParentIssueKey"    //
+	StatusNameKey     = "StatusNameKey"     //
+	IssueForceOpenKey = "IssueForceOpenKey" //
 )