+++ pkg/status/transition.go
@@ -1,0 +1,9 @@
+package status
+
+type DoTransitionObject struct {
+	Transition InnerTransition `json:"transition"`
+}
+
+type InnerTransition struct {
+	Id string `json:"id"`
+}